# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AutorizacionCarga": {
        "description": "Permiso temporal para escribir un blob concreto.\n\nEl archivo se sube directamente al almacenamiento con esta URL. No\natraviesa la puerta de enlace ni los servicios de aplicación.",
        "properties": {
          "encabezados": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Encabezados obligatorios de la solicitud de subida",
            "title": "Encabezados",
            "type": "object"
          },
          "expira": {
            "format": "date-time",
            "title": "Expira",
            "type": "string"
          },
          "metodo": {
            "const": "PUT",
            "default": "PUT",
            "title": "Metodo",
            "type": "string"
          },
          "ruta_blob": {
            "description": "Ruta que debe confirmarse tras la subida",
            "title": "Ruta Blob",
            "type": "string"
          },
          "url": {
            "description": "URL con firma de acceso compartido",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "ruta_blob",
          "expira"
        ],
        "title": "AutorizacionCarga",
        "type": "object"
      },
      "ConfirmacionCarga": {
        "properties": {
          "ruta_blob": {
            "title": "Ruta Blob",
            "type": "string"
          }
        },
        "required": [
          "ruta_blob"
        ],
        "title": "ConfirmacionCarga",
        "type": "object"
      },
...
      "EntradaHistorial": {
        "properties": {
          "estado": {
            "$ref": "#/components/schemas/Estado"
          },
          "id_caso": {
            "title": "Id Caso",
            "type": "string"
          },
          "id_corrida": {
            "title": "Id Corrida",
            "type": "string"
          },
          "marca_tiempo": {
            "format": "date-time",
            "title": "Marca Tiempo",
            "type": "string"
          },
          "nombre_caso": {
            "title": "Nombre Caso",
            "type": "string"
          },
          "npv_musd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Npv Musd"
          },
          "usuario": {
            "title": "Usuario",
            "type": "string"
          },
          "version_motor": {
            "title": "Version Motor",
            "type": "string"
          }
        },
        "required": [
          "id_corrida",
          "id_caso",
          "nombre_caso",
          "estado",
          "usuario",
          "marca_tiempo",
          "version_motor"
        ],
        "title": "EntradaHistorial",
        "type": "object"
      },
      "Estado": {
        "enum": [
          "borrador",
          "calculada",
          "congelada"
        ],
        "title": "Estado",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaginaHistorial": {
        "properties": {
          "continuacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Continuacion"
          },
          "corridas": {
            "items": {
              "$ref": "#/components/schemas/EntradaHistorial"
            },
            "title": "Corridas",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "corridas",
          "total"
        ],
        "title": "PaginaHistorial",
        "type": "object"
      },
...
      "Problema": {
        "description": "Cuerpo de error. Explica qué pasó y cómo resolverlo.",
        "properties": {
          "detalle": {
            "title": "Detalle",
            "type": "string"
          },
          "restriccion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Restricción del cuadro de control que bloquea, si aplica",
            "title": "Restriccion"
          }
        },
        "required": [
          "detalle"
        ],
        "title": "Problema",
        "type": "object"
      },
...
      "ResultadoValidacion": {
        "description": "Validación de una plantilla antes de admitirla al cálculo.",
        "properties": {
          "filas_leidas": {
            "title": "Filas Leidas",
            "type": "integer"
          },
          "hallazgos": {
            "description": "Problemas encontrados, con hoja y celda cuando aplica",
            "items": {
              "type": "string"
            },
            "title": "Hallazgos",
            "type": "array"
          },
          "huella": {
            "default": "",
            "description": "SHA-256 del contenido admitido",
            "title": "Huella",
            "type": "string"
          },
          "valida": {
            "title": "Valida",
            "type": "boolean"
          }
        },
        "required": [
          "valida",
          "filas_leidas"
        ],
        "title": "ResultadoValidacion",
        "type": "object"
      },
...
      "SolicitudCarga": {
        "description": "Petición de autorización para subir una plantilla.",
        "properties": {
          "nombre_archivo": {
            "description": "Nombre original. Se usa para validar la extensión, no para la ruta",
            "minLength": 5,
            "title": "Nombre Archivo",
            "type": "string"
          },
          "tamanio_bytes": {
            "exclusiveMinimum": 0.0,
            "maximum": 26214400.0,
            "title": "Tamanio Bytes",
            "type": "integer"
          },
          "tipo": {
            "enum": [
              "produccion",
              "capex",
              "opex"
            ],
            "title": "Tipo",
            "type": "string"
          }
        },
        "required": [
          "tipo",
          "nombre_archivo",
          "tamanio_bytes"
        ],
        "title": "SolicitudCarga",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "name": "INVA · Servicio INVA-01-2026-182"
    },
    "description": "\nServicios de la Plataforma de Evaluación Económica · Modelo MINSUR.\n\nReproduce la lógica del modelo económico corporativo para que el área de\nProyectos evalúe escenarios de forma autónoma, con trazabilidad completa.\n\n**Autenticación.** Token de Microsoft Entra ID del tenant corporativo. El\nperfil se deriva de la reclamación `groups`; la plataforma no consulta la\nmembresía a Microsoft Graph.\n\n**Archivos.** Las plantillas se cargan y los reportes se descargan\ndirectamente contra el almacenamiento, mediante firmas de acceso compartido\nde corta vigencia. No atraviesan esta API.\n\n**Respuestas 501.** Indican una operación cuya implementación depende de un\ninsumo del cliente aún no recibido. El cuerpo incluye la restricción del\ncuadro de control que la bloquea.\n",
    "title": "Plataforma de Evaluación Económica · MINSUR",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auditoria": {
      "get": {
        "description": "Bitacora de solo escritura, restringida al perfil Auditor.\n\nCada registro contiene el identificador de la corrida, el usuario\nresponsable, la marca de tiempo, la accion y, ante modificacion de\ninsumos o parametros, los valores anterior y posterior.\n\nLas entradas van encadenadas: cada una incluye el resumen de la anterior,\nde modo que suprimir una intermedia rompe la cadena y se detecta.",
        "operationId": "auditoria_api_auditoria_get",
        "parameters": [
          {
            "in": "query",
            "name": "caso",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Caso"
            }
          },
          {
            "in": "query",
            "name": "corrida",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Corrida"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Auditoria Api Auditoria Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Bitacora de auditoria",
        "tags": [
          "Historial"
        ]
      }
    },
...
    "/api/casos/{id_caso}/plantillas/autorizacion": {
      "post": {
        "description": "Emite una firma de acceso compartido acotada a un solo archivo.\n\nCinco restricciones sostienen la seguridad del mecanismo:\n\n- Firma de **delegación de usuario**, no clave de cuenta. El\n  almacenamiento tiene deshabilitado el acceso por clave compartida: no\n  existe clave que firmar.\n- Vigencia de quince minutos.\n- Permiso acotado a **un blob concreto**, no al contenedor.\n- Escritura y creación, sin lectura: quien carga no puede leer lo que ya\n  existe.\n- **El nombre del blob lo determina el servidor.** Aceptar una ruta del\n  cliente permitiría escribir sobre la plantilla de otra evaluación.",
        "operationId": "autorizar_carga_api_casos__id_caso__plantillas_autorizacion_post",
        "parameters": [
          {
            "in": "path",
            "name": "id_caso",
            "required": true,
            "schema": {
              "examples": [
                "CASO-SR-2026-014"
              ],
              "title": "Id Caso",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudCarga"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutorizacionCarga"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Autorizar la carga directa de una plantilla",
        "tags": [
          "Plantillas"
        ]
      }
    },
    "/api/casos/{id_caso}/plantillas/confirmar": {
      "post": {
        "description": "Comprueba que el archivo llegó y lo valida antes de admitirlo.\n\nLa validación de estructura, unidades, clasificaciones y horizonte\ntemporal depende de las plantillas definitivas que elabora MINSUR\n(`R-07`). INVA implementa su lectura y validación, no su diseño.",
        "operationId": "confirmar_carga_api_casos__id_caso__plantillas_confirmar_post",
        "parameters": [
          {
            "in": "path",
            "name": "id_caso",
            "required": true,
            "schema": {
              "examples": [
                "CASO-SR-2026-014"
              ],
              "title": "Id Caso",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmacionCarga"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoValidacion"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Confirmar y validar una plantilla cargada",
        "tags": [
          "Plantillas"
        ]
      }
    },
...
    "/api/historial": {
      "get": {
        "description": "Corridas ordenadas de la mas reciente a la mas antigua.\n\nCada entrada incluye la version del motor con la que se calculo, no la\nvigente: es lo que permite entender por que dos corridas del mismo caso\ndan distinto.\n\nEs la quinta de las cinco comprobaciones obligatorias del pase.",
        "operationId": "historial_api_historial_get",
        "parameters": [
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 200,
              "minimum": 1,
              "title": "Limite",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "caso",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Caso"
            }
          },
          {
            "in": "query",
            "name": "continuacion",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Continuacion"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaHistorial"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Historial de evaluaciones",
        "tags": [
          "Historial"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "A través de API Management",
      "url": "/"
    }
  ]
}